requirements.txt
numpy
pandas
matplotlib
seaborn
pandas_gbq
scikit-learn>=1.4
xgboost
lightgbm
catboost
statsmodels
pmdarima
pyaml
mlflow
shap
nbformat>=2.4.0
nbconvert
ipywidgets